+++ benchmark/performance/arithmetic/native/performance.cpp
@@ -16,7 +16,7 @@
 #include <universal/verification/performance_runner.hpp>
 
 template<typename NativeFloat>
-void CopyWorkload(uint64_t NR_OPS) {
+void CopyWorkload(size_t NR_OPS) {
 	using namespace sw::universal;
 	NativeFloat a,b,c;
